Add second NFD CR update to overcome cluster-nfd-operator bug.
The cluster-nfd-operator returns after update the NFD CR's finalizer,
expecting to have an additional reconciliation request that would move
forward with the update of the rest of the NFD CR's components.

The predicates used for the NFD CR watches, exclude any update events
that do not increment the object's generation. Updating the finalizers of an
object do not increment the generation, leading to the NFD CR to not be
reconciled after getting created.

Resources:
- https://github.com/openshift/cluster-nfd-operator/blob/master/controllers/nodefeaturediscovery_controller.go#L379
- https://github.com/openshift/cluster-nfd-operator/blob/master/controllers/nodefeaturediscovery_controller.go#L490
- https://github.com/openshift/cluster-nfd-operator/blob/c3b05399428719230aa220c5dcad7235733ca950/controllers/nodefeaturediscovery_finalizers.go#L84

Signed-off-by: Michail Resvanis <mresvani@redhat.com>