Allow namespace to run privileged pods (#32)
With OpenShift 4.11, the [Pod Security
admission](https://kubernetes.io/docs/concepts/security/pod-security-admission/)
will be enabled and warning will be issued for pods that do not comply.
This will generate noise and the fix is fairly simple.

The GPU Operator requires privileged pods and we deploy it in the GPU
Add-on namespace, so we can simply update the GPU Add-on namespace
labels to allow privileged pods.

Signed-off-by: Fabien Dupont <fabiendupont@pm.me>