Add app label to prometheus CRs and use it as a selector (#56)
The Prometheus CR needs LabelSelector to add `{Pod,Service}Monitor`s
and `Rule`s in its configuration. `metav1LabelSelector`s do not support
`OR`-ed expressions, only `AND`-ed ones, which is the reason to have a
common label among all prometheus CRs. Then we simply check for this
label's existence. This change is driven by the fact that the GPU
operator adds the `app` label to all of its prometheus-operator CRs.

Signed-off-by: Michail Resvanis <mresvani@redhat.com>